Restructure the make-dictionaries-readable source files into the clp namespace.
