+++ components/core/src/clp/make_dictionaries_readable/README.md
